requirements.txt
requests==2.20.1
pycrypto==2.6.1
protobuf==3.6.1
